refactor(api) migrate serializer
