+++ api/tests/prescription/organization-learner/unit/infrastructure/serializers/jsonapi/organization-learner-identity-serializer_test.js
@@ -1,6 +1,6 @@
-import { OrganizationLearner } from '../../../../../lib/domain/models/OrganizationLearner.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/organization-learner-identity-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import { OrganizationLearner } from '../../../../../../../lib/domain/models/OrganizationLearner.js';
+import * as serializer from '../../../../../../../src/prescription/organization-learner/infrastructure/serializers/jsonapi/organization-learner-identity-serializer.js';
+import { expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | organization-learner-identity-serializer', function () {
   describe('#serialize', function () {
